Move to_pandas tests to own module
